Default to using libunwind-ptrace for native unwinding
After running another round of testing, there doesn't seem to be any
performance benefit on using gimli based unwinder over libunwind.
Since this complicates the code significantly and we're still having
to use both, remove the gimli unwinder in favour of libunwind.
