+++ remoteprocess/Cargo.toml
@@ -10,7 +10,6 @@ license = "GPL-3.0"
 build="build.rs"
 
 [dependencies]
-gimli = "0.19.0"
 libc = "0.2"
 log = "0.4"
 proc-maps = "0.1.6"
