+++ remoteprocess/src/osx/mod.rs
@@ -1,17 +1,11 @@
-#[doc(hidden)]
-pub mod compact_unwind;
 mod utils;
-#[cfg(unwind)]
-mod symbolication;
 mod mach_thread_bindings;
-#[cfg(unwwind)]
-mod unwinder;
 
 use std;
 use std::convert::TryInto;
 use mach;
 
-use super::{ProcessMemory, Error};
+use super::Error;
 use mach::kern_return::{KERN_SUCCESS};
 use mach::port::{mach_port_name_t, MACH_PORT_NULL};
 use mach::traps::{task_for_pid, mach_task_self};
@@ -26,8 +20,6 @@ use mach::thread_status::x86_THREAD_STATE64;
 use mach::thread_act::{thread_get_state};
 
 pub use self::utils::{TaskLock, ThreadLock};
-#[cfg(unwwind)]
-pub use self::unwinder::Unwinder;
 
 use libproc::libproc::proc_pid::{pidpath, pidinfo, PIDInfo, PidInfoFlavor};
 
@@ -122,11 +114,6 @@ impl Process {
             ret.push(Thread{tid});
         }
         Ok(ret)
-    }
-
-    #[cfg(unwind)]
-    pub fn unwinder(&self) -> Result<Unwinder, Error> {
-        Ok(Unwinder::new(self.pid)?)
     }
 }
